+++ src/App.vue
@@ -35,7 +35,7 @@ const init = async () => {
   try {
     useNotifications()
     await web3Store.detectProviders()
-
+    await networkStore.loadNetworks()
     await web3Store.init(networkStore.list.map(el => el.chain_id))
 
     document.title = config.APP_NAME