+++ src/common/network-switcher/NetworkSwitcher.vue
@@ -1,6 +1,6 @@
 <template>
   <drop-down
-    v-if="provider.selectedAddress"
+    v-if="provider.address"
     :class="networkSwitcherClasses"
     :right="dropDownShift"
     :disabled="isSwitchingChain"
@@ -40,7 +40,7 @@ import { ref, computed, onMounted } from 'vue'
 import { DropDown, NetworkItem, Loader, ErrorMessage } from '@/common'
 import { useWeb3ProvidersStore, useNetworksStore } from '@/store'
 import { ErrorHandler, getNetworkScheme, switchNetwork } from '@/helpers'
-import { ChainId } from '@/types'
+import { ChainId } from '@distributedlab/w3p'
 import { useWindowSize } from '@vueuse/core'
 import { WINDOW_BREAKPOINTS } from '@/enums'