+++ src/forms/purchase-book-form/integrated-payments/RarimoFlow.vue
@@ -247,7 +247,7 @@ const initializeCheckout = async (
   sourceChain: BridgeChain,
   targetChain: BridgeChain,
 ) => {
-  if (!tokenPrice.value || !provider.value.selectedAddress) return
+  if (!tokenPrice.value || !provider.value.address) return
 
   const nftPrice = new BN(book.pricePerOneToken, {
     decimals: tokenPrice.value?.token.decimals,
@@ -257,7 +257,7 @@ const initializeCheckout = async (
     .toFixed()
 
   await initCheckout(sourceChain, targetChain, {
-    recipient: provider.value.selectedAddress,
+    recipient: provider.value.address,
     nftPrice,
   })
 }
@@ -265,7 +265,7 @@ const initializeCheckout = async (
 const submitFunc = async (editorInstance: UseImageEditor | null) => {
   if (
     !priceRaw ||
-    !provider.value.selectedAddress ||
+    !provider.value.address ||
     !editorInstance ||
     !tokenPrice.value
   )
@@ -280,8 +280,8 @@ const submitFunc = async (editorInstance: UseImageEditor | null) => {
     const { buyParams, signature } = await buildFormMintData({
       banner,
       book,
-      account: provider.value.selectedAddress,
-      chainId: Number(provider.value.chainId),
+      account: provider.value.address,
+      chainId: Number(config.DEFAULT_CHAIN_ID),
       tokenAddress: '',
     })
 
@@ -356,7 +356,7 @@ watch(
 )
 
 watch(
-  () => provider.value.selectedAddress,
+  () => provider.value.address,
   async () => {
     isLoading.value = true
     await initForm()