+++ src/pages/my-nfts/MyNftItemPage.vue
@@ -54,20 +54,13 @@ import {
 } from '@/common'
 
 import { ErrorHandler } from '@/helpers'
-import { ref, computed } from 'vue'
-import { useNftTokens, useErc721 } from '@/composables'
+import { ref } from 'vue'
+import { useNftTokens } from '@/composables'
 import { useI18n } from 'vue-i18n'
-import { useWeb3ProvidersStore } from '@/store'
-import { router } from '@/router'
-import { ROUTE_NAMES } from '@/enums'
 import { TokenFullInfo } from '@/types'
 
 const { t } = useI18n()
 const { getNft } = useNftTokens()
-const { init: initErc721, getOwner } = useErc721()
-
-const web3Store = useWeb3ProvidersStore()
-const provider = computed(() => web3Store.provider)
 
 const TABS = {
   myPurchase: {
@@ -94,19 +87,11 @@ const nftToken = ref<TokenFullInfo | undefined>()
 const init = async () => {
   try {
     const data = await getNft(props.contractAddress, props.id)
-    initErc721(data.tokenContract)
-
-    const owner = await getOwner(data.tokenId)
-
-    // not showing this page if its not owner of nft
-    if (provider.value.selectedAddress !== owner) {
-      router.push({ name: ROUTE_NAMES.myNfts })
-      return
-    }
 
     nftToken.value = data
   } catch (error) {
     ErrorHandler.processWithoutFeedback(error)
+
     if (error instanceof Error) errorMsg.value = error.message
   }
   isLoaded.value = true